requirements.txt
pytorch_lightning~=1.6.4
torch~=1.10.1
torchmetrics~=0.9.0
torchvision~=0.11.2
PyYAML~=6.0
pandas~=1.4.2
seaborn~=0.11.2